Extract `ShCmd` execution logic to `ShCmdExecutor`.
